First step in transitioning the current decompression code to a more generic system of stackable stream transforms. In particular, I believe I've edited every place that called the decompressor directly to go through new __archive_read_ahead(), __archive_read_consume() and __archive_read_skip() functions, so the rest of the work here will not require any changes to the format handlers.
I've also laid out the new types that will be needed for this.
Next step is to rewrite the decompressors to the new interface,
and overhaul the decompression auction to juggle generic "sources."
Then I'll be able to consolidate reblocking into a single place;
the transforms can emit arbitrary-sized blocks and the current
decompress_none logic will be used to reblock as needed by the
consumer format.

The initial impetus for this was to simplify the decompressors by
consolidating the reblocking logic.  I recently came up with
some other transforms I'd like to implement (including new
decompressors, an encryption filter for secure backup, and
uudecode handling to simplify test harnesses) that would also
benefit from this.  Eventually, I think I might be able to
standardize the interface for new transforms enough to allow
libarchive clients to register their own transforms complete
with bidding logic (the old interface was too wired into libarchive
internals for the API to be exported).  In the very long term,
this might divorce the transformation logic from the rest of
libarchive enough to allow it to be packaged as an independent
library.

SVN-Revision: 234