In archive_write_disk:  If archive_write_header() fails to create the file, that's a failure and should return ARCHIVE_FAILED. Metadata restore failures still return ARCHIVE_WARN, because that's non-critical.  Fix test_write_disk_secure test to verify the correct return code in one case; add test_write_disk_failures to do another very simple test of restore failure.
In particular, this fixes the problem found with xattr restores
that a failure in setting an xattr would prevent the file and other
metadata from being written.  Now, an xattr restore failure generates
a warning but does not prevent the rest of that file's data and metadata
from being properly written.

SVN-Revision: 419