Check in getdate.c and getdate.h.
On systems that have Yacc or Bison, these are constructed from
getdate.y.  However, as we push the portability envelope, we're
encountering more systems that don't have Yacc/Bison and the
lack of these files is making it difficult for people on those
platforms to contribute.

Eventually, I'd like to rewrite getdate() into straight C
and lose the Yacc dependency entirely, but I'm not quite
ready for that.

SVN-Revision: 689