Testing:  test_pathmatch.c now has 100% test coverage of pathmatch.c Allow '$' to force anchor at end iff PATHMATCH_NO_ANCHOR_END is in flags. Fix backslash escaping so that patterns such as "\\/./" or "/\\./" match an explicit "/./" but do not match plain "/", even though "/./" does ordinarily match "/".
I think this is finally sufficient to match my long-standing desire
to be able to extract "./foo/" by just asking for "foo".
There's still an open question whether this can be extended to handle
Windows-style \ path separators without causing conflicts with
the use of \ as a quote character.  I don't think it can.

SVN-Revision: 824