Rework the style of CHECK_FUNCS() to better match the style of built-in cmake tests:  * One test per invocation  * macro name is explicit
Convert some of the tests to CHECK_SYMBOL_EXISTS(),
which further reduces some of the magic within the
old CHECK_FUNCS.  It is now primarily an extended
version of CHECK_FUNCTION_EXISTS that understands the
glibc __stub_ convention for marking functions that
should be treated as nonexistent even though they
do exist in the local library.  I've renamed the
macro to CHECK_FUNCTION_EXISTS_GLIBC to better
reflect this.

SVN-Revision: 1308