Simplify try-compile for SIZE_MAX
This constant may be defined in one of a few headers that we always
include anyway.  It is simpler and more reliable to look for the
constant in all the possible headers at once.  Otherwise we would need
to store the result for each header in a separate cache entry.

SVN-Revision: 1655