Use ADD_TEST_28 macro to fix tests with the VS IDE
This macro simplifies creation of tests that refer to executables built
in the project.  We can pass executables as arguments to a test using a
CMake 2.8 "generator expression" of the form "$<TARGET_FILE:myexe>".
The macro translates the arguments into something CMake 2.6 understands.

This change also fixes the tests so that they run correctly in build
trees created for multi-configuration build tools like the VS IDE.
These tools place executables in a per-configuration subdirectory so
references to executables in test arguments must use a test-time
substitution.  The ADD_TEST_28 macro handles this automatically.

SVN-Revision: 1672