Support openssl without SHA-2 hash functions
Previously we assumed that <openssl/sha.h> provides all the SHA-2 hash
functions if it exists.  Some OpenSSL installations do not, so we need
to test for them at configuration time.  Currently we perform the tests
only for CMake builds.  We preserve the old behavior for the autotools
build by hard-coding availability, and leave a TODO comment for someone
to actually add the checks.

SVN-Revision: 1691