Remove STDC_HEADERS check from non-autoconf builds
This autoconf check is used only within its configure tests and is not
used in libarchive source code.  The CMake configuration test used a
try_run(), which prevents cross-compiling without manual specification
of the result.  Since the result of the test is not used anyway we
simply remove it.

SVN-Revision: 1692