Adjust this test slightly so it doesn't depend on a particular order of traversing the directory.
In particular, this test almost passes against
GNU tar 1.23.  The difference being that GNU tar
considers it an error when a hardlink cannot
be extracted because the target was erased by
--strip-components and bsdtar does not consider
this an error.  Both extract the same files;
the only difference is that GNU tar emits
some error messages and exits with an error
status, while bsdtar exits with success.

SVN-Revision: 2224