Issue 83:  Directory timestamps truncated to microsecond resolution.
This patch reworks the time-restore logic to use the same
system calls in both the end-of-archive cleanup pass and
the regular file restore case.  This ensures consistent
timestamp restores for both cases.  It also isolates the
BSD birthtime support into one place so it will be consistently
applied.

Submitted by:	a.bocaniciu  computer.org

SVN-Revision: 2231