If opening the filter pipeline fails, immediately close the pipeline and return with an error without initializing the format writer.  Ensure that close requests propagate up the pipeline even if a particular stage hasn't registered a closer.
In particular, this fixes a crash that occurs when
write openers would fail but the format would get
initialized anyway.  At close time, formats such
as tar perform a write that would crash because the
pipeline wasn't fully initialized.  With this
change, the format never gets initialized in
this case so it doesn't need to be cleaned up.

SVN-Revision: 2539