If we see junk when we're expecting a 'PK' signature block, scan forward to see if we can find a suitable signature.
This is necessary to read archives that have been modified
by some Zip utilities that update entries in-place without
compacting the entire archive.

Of course, this would be very natural if libarchive
used the Central directory.  But even when libarchive
does support the Central directory, this kind of logic
will still be useful for reading Zip archives in streaming
mode.

SVN-Revision: 2781