First part of the NFS4 ACL support.
This renames a few things to acknowledge that there really
is more than one kind of ACL in the world and extends
the basic ACL storage to support the NFS4/NTFS ACL
mode bits.  The ACL storage has also gained some
error checks to ensure that a single ACL does not have
both NFS4/NTFS and POSIX.1e ACEs.

SVN-Revision: 2813