Introudce "tar:utf8type=libarchive2x" option for the incorrect UTF-8 string which libarchive 2.x makes in wrong assumption about wchar_t. The option works only for pax format.
Currently libarchive 3 correctly translates UTF-8 string from/to current locale
string, but we cannot accordingly handle the incorrect UTF-8 on the some
platforms wchar_t of which is not Unicode and users are not using UTF-8 locale.
So we should support the UTF-8 string to be properly translated to current
locale string.

SVN-Revision: 3254