Mask SIGPIPE for libarchive test, tar, and cpio.
As reported at:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=641265

The SIGPIPE occurs when feeding data through a decompression
program; the external program may see an end-of-data marker
and exit before we've finished feeding it all of the data.
(This happens when the data has garbage padding at the end.)
We want to catch and ignore the EPIPE since we can still
read the data coming out of the decompression program.

SVN-Revision: 3751