Issue 190:  Fix build when liblzma is installed through MacPorts.
The probe for liblzma finds it in /opt/local, which adds
/opt/local/include to the include search path, which causes
iconv.h to be found there at compile time, which breaks because
libiconv was found in /usr at configure time.

This change causes cmake to prefer /opt/local for everything,
so that libiconv is pulled from /opt/local, and we avoid
the breakage from iconv mismatch.

I have a nasty suspicion that similar problems can happen
with many package systems.

SVN-Revision: 3753