Issue 224:Mishandling CD9660 images with RockRidge extensions from FreeBSD makefs.
A file size of the symblic link file in ISO images made by makefs is not zero and
its location is not negative. That makes ISO reader misleading that that file is
a hard like file.
We have to check if the file type is symbolic like or not, if so and its file size
is not zero, reset the file size to zero and also reset the location to -1.

SVN-Revision: 4181