Issue 225: Misreading directories in MSDOS Zip files.
These files don't put a POSIX "mode" value in the "external
attributes" field.  We used to always assume a regular file if we
didn't see that.

Now, the code leaves the mode set to zero in that case.  When we read
the local file header, the fallback heuristics will kick in and
examine the filename to determine whether this is a directory
or a regular file.

SVN-Revision: 4190