Refactored common code into a single function
The tar_atol8() and tar_atol10() functions were essentially
identical. The code has been refactored to a common function
with tar_atol8() and tar_atol10() calling the commont function
with the radix instead duplicating the code.

The logic was also changed slightly to prevent a pointer
dereference when a zero length character array is passed to
the conversion routine.