Add CMake option to explicitly enable/disable /SAFESEH linker flag for Visual Studio builds.
Visual Studio 12 enables this option by default. Enabling /SAFESEH will produce build failures when
attempting to link other libraries which were not built with /SAFESEH. This would typically be
the case for libraries built using the mingw toolchain, where the option to use structured exception
handling is not yet available.

The option is used as follows.
  Setting to "YES" sets "/SAFESEH" linker flag
  Setting to "NO" sets "/SAFESEH:NO" linker flag
  Setting to any other value such as "AUTO" won't do anything (set the default linker flag from Visual Studio)