Better sanity check in case of multiple EOCD files.
It may happen that a file has multiple EOCD headers in the last 16k of the
archive. In such case, the first one would be chosen.

This can be simply avoided by adding a sanity check, that the entire rest of
the file is in the EOCD comment, which is a valid assumption according to the
ZIP file format specification.

TEST=Tested manually with such ZIP archive.
BUG=None