Improve readability of ternary expression checking for empty string
In expressions of the form

    m != NULL && m[0] == '\0' ? NULL : m

the goal is to get a `NULL` pointer if `m` is either `NULL` or an empty
string.  This is not clear because in the `m == NULL` case we use the
`m` side to get `NULL` instead of an explicit `NULL`.  Clarify the
intent by using the form

    (m != NULL && m[0] != '\0') ? m : NULL;

instead.