Add a "moreEntries" flag to indicate that the current directory still contains entries that need to be reported to caller.
Do not use t->d value for this anymore: when sorting entries, we have to parse
the entire directory before the sort, which means that tree_dir_iterate() will
close the directory before having the first entry returned to the caller.
As such, split the logic behind directory opening/closing (controlled by
t->d value) and entries reporting to the caller (controlled by the
"moreEntries" flag).