As requested by Tim, always perform the sort when we are below a "small" number of entries (32). Caller can override this maximum and ask for a complete sort through ARCHIVE_READDISK_SORT_ENTRIES flag (the maximum will then be set to SIZE_MAX).
Current implementation now stores at least a "small" number of dirent structures
inside the array, and performs sort when we are below the maximum. Otherwise,
when we are above maximum (or an error occured), the entries will first be
obtained from the array, then from the following readdir() calls (through
tree_dir_iterate()).

XXX As optimization, the default maximum (32) of dirent structures should always
be kept pre-allocated, as we will always cache that many structures.
This comes in the next patch, as dirent structures size might vary during tree
traversal (array entries should therefore be adapted to it).