Replace tree traversal logic from a LIFO to a FIFO.
The old implementation uses a LIFO logic (very close to a work stack),
which has a shortcoming when we are sorting: when entries are pushed
onto the work stack, the last element of the sort will be popped first,
leading to a reversed order traversal when walking down the tree.

This patch implements a work queue instead. Elements are pushed into
a queue, therefore preserving the order in which they were submitted by
caller.

Implementation details:

When caller notifies libarchive that it has to descend into a directory,
it pushes it into a queue that contains all the pushed children from current
directory. The first pushed child is then descended into for iteration.

Once done with current entry iteration, libarchive proceeds with the
next queued entry. It may encounter the following situations:
- the next entry to process is a child of the same parent (e.g. the two
   entries are positioned at the same depth within the tree). It
   will ascend back to parent, and descend into the next child;
- the next entry to process is a child of the current directory (e.g. the
   next entry is one-level below the current entry). libarchive
   will therefore descend directly into it as it does not have to ascend
   back to the parent;
- the next entry to process is one level above the child, which means
   that we reached the end of current's subtree. It will have to ascend
   back to parent, and queue parent as next entry as we finished traversing
its subtree.

Notable changes:

- tree_descend() and tree_ascend() functions now deal with tree's state
  update, including full path, basename and file descriptors;
- implemented a tree_update_basename() functions that updates basename
  of t->path directly;
- tree_entry allocation functions (tree_push, tree_new_entry,
  tree_first_entry) return a pointer; caller can deal with allocations
  errors more easily;
- tree_next() flags are now exclusive: a directory can be in either
  a "needsDescent", "needsIterate" or "needsAscent" state, which respectively
  represent a directory that shall be "descended into", "iterated over
  its entries", or "ascended from". Two of these states map directly to
  visit types:
    - TREE_POSTDESCENT (needsDescent);
    - TREE_POSTASCENT (needsAscent).
  The "needsIterate" state corresponds to the first visit of the directory
  after descending into it and will return all TREE_REGULAR entries
  found within.

No regression observed when running the test suite.