Move the pattern handling code inside tree_dir_iterate(), as this function is the main entry point for iterating over directories now.
Add a safety check so it cannot be used to open up directories
while no pattern has been specified.

Handle "ERROR_NO_MORE_FILES" differently from other errors as it
indicates that we are done iterating (which is not an error per see,
while other kinds of errors must be returned back to caller).