Issue #748: Zip decompression failure with highly-compressed data
Previously, we stopped driving the decompressor as soon as
we hit end of input, but in some cases, the decompressor
has internal state that can continue generating output
even when there is no more input.  So we now feed zero-length
blocks into the decompressor until the decompressor tells us
it is finished.