Fix a rar crash
On certain rars, I can get an "Invalid symbol" error, and a segfault
immediately afterwards. The error is caused by Ppmd7_DecodeSymbol()
returning an error. If the API user tries to continue decompression
after the error, Ppmd7_DecodeSymbol is called again then the
Ppmd7_GetStats call (immediately following the code touched by the
patch) gets an invalid pointer. Work around the cause of the crash. I'm
not sure about the deeper reason for this.